+++ src/main/java/dev/petshopsoftware/utilities/Util/InputChecker/FloatChecker.java
@@ -14,19 +14,19 @@ public class FloatChecker extends InputChecker<Float> {
 		}
 	}
 
-	public FloatChecker min(int length, String message) {
+	public FloatChecker min(float length, String message) {
 		return (FloatChecker) check(input -> input >= length, message);
 	}
 
-	public FloatChecker max(int length, String message) {
+	public FloatChecker max(float length, String message) {
 		return (FloatChecker) check(input -> input <= length, message);
 	}
 
-	public FloatChecker range(int min, int max, String message) {
+	public FloatChecker range(float min, float max, String message) {
 		return (FloatChecker) check(input -> input >= min && input <= max, message);
 	}
 
-	public FloatChecker mod(int mod, String message) {
+	public FloatChecker mod(float mod, String message) {
 		return (FloatChecker) check(input -> input % mod == 0, message);
 	}
 }
